fix: split Terminal into modular files, remove duplicate declarations
